add sysfs initial implementation of iccom sk if
Added initial implementation for the iccom_sockets_if driver. Now it
encompasses the device creation, the protocol family configuration and
the linkage of an iccom device to an iccom_sockets device. Messages from
and to the iccom layer are flowing.
The test script automatically creates, changes the protocol family, links
an iccom_device to an iccom_socket device and reads the transport layer
to check the message sent from iccom_sockets_if layer.
Added a new README_ICCOM_SK_IF for guiding the user to use the ICCOM sk
if through the API.

RTC1-1659227
RTC1-1660659