add device link and delete functionality
Added the linkage between an iccom sk device and an iccom device
in order to stablish a hierarchical responsibility between devices
in different layers. Therefore, an iccom device cannot be remove if
it has an iccom sk device linked to it.
Moreover, the delete functionality is now exposed to the user space
through sysfs for the iccom sk devices.

RTC1-1698218